All child classes of Order are declared within Order.h and defined within Order.cpp. A free function named testOrdersLists() was added to OrdersDriver.cpp and main() within Test.cpp was modified accordingly.
